docs: align schema bigint foreign keys

+++ docs/schema.sql
@@ -43,7 +43,7 @@ CREATE TYPE order_status AS ENUM (
 -- =========================
 
 CREATE TABLE users (
-    user_id SERIAL PRIMARY KEY,
+    user_id BIGSERIAL PRIMARY KEY,
     full_name VARCHAR(100) NOT NULL,
     email VARCHAR(100) UNIQUE NOT NULL,
     password_hash VARCHAR(255) NOT NULL,
@@ -57,7 +57,7 @@ CREATE TABLE users (
 -- =========================
 
 CREATE TABLE categories (
-    category_id SERIAL PRIMARY KEY,
+    category_id BIGSERIAL PRIMARY KEY,
     category_name VARCHAR(50) UNIQUE NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
@@ -67,8 +67,8 @@ CREATE TABLE categories (
 -- =========================
 
 CREATE TABLE products (
-    product_id SERIAL PRIMARY KEY,
-    category_id INT NOT NULL REFERENCES categories(category_id) ON DELETE CASCADE,
+    product_id BIGSERIAL PRIMARY KEY,
+    category_id BIGINT NOT NULL REFERENCES categories(category_id) ON DELETE CASCADE,
     name VARCHAR(100) NOT NULL,
     description TEXT,
     price DECIMAL(10, 2) NOT NULL CHECK (price >= 0),
@@ -83,8 +83,8 @@ CREATE TABLE products (
 -- =========================
 
 CREATE TABLE inventory (
-    inventory_id SERIAL PRIMARY KEY,
-    product_id INT UNIQUE NOT NULL REFERENCES products(product_id) ON DELETE CASCADE,
+    inventory_id BIGSERIAL PRIMARY KEY,
+    product_id BIGINT UNIQUE NOT NULL REFERENCES products(product_id) ON DELETE CASCADE,
     quantity INT NOT NULL DEFAULT 0 CHECK (quantity >= 0),
     last_updated TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
@@ -94,16 +94,16 @@ CREATE TABLE inventory (
 -- =========================
 
 CREATE TABLE carts (
-    cart_id SERIAL PRIMARY KEY,
-    user_id INT UNIQUE NOT NULL REFERENCES users(user_id) ON DELETE CASCADE,
+    cart_id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT UNIQUE NOT NULL REFERENCES users(user_id) ON DELETE CASCADE,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE cart_items (
-    cart_item_id SERIAL PRIMARY KEY,
-    cart_id INT NOT NULL REFERENCES carts(cart_id) ON DELETE CASCADE,
-    product_id INT NOT NULL REFERENCES products(product_id),
+    cart_item_id BIGSERIAL PRIMARY KEY,
+    cart_id BIGINT NOT NULL REFERENCES carts(cart_id) ON DELETE CASCADE,
+    product_id BIGINT NOT NULL REFERENCES products(product_id),
     quantity INT NOT NULL CHECK (quantity > 0),
     unit_price DECIMAL(10, 2) NOT NULL CHECK (unit_price >= 0),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -116,8 +116,8 @@ CREATE TABLE cart_items (
 -- =========================
 
 CREATE TABLE orders (
-    order_id SERIAL PRIMARY KEY,
-    user_id INT NOT NULL REFERENCES users(user_id) ON DELETE CASCADE,
+    order_id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT NOT NULL REFERENCES users(user_id) ON DELETE CASCADE,
     order_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     status order_status NOT NULL DEFAULT 'PENDING',
     total_amount DECIMAL(12, 2) NOT NULL CHECK (total_amount >= 0)
@@ -128,9 +128,9 @@ CREATE TABLE orders (
 -- =========================
 
 CREATE TABLE order_items (
-    order_item_id SERIAL PRIMARY KEY,
-    order_id INT NOT NULL REFERENCES orders(order_id) ON DELETE CASCADE,
-    product_id INT NOT NULL REFERENCES products(product_id),
+    order_item_id BIGSERIAL PRIMARY KEY,
+    order_id BIGINT NOT NULL REFERENCES orders(order_id) ON DELETE CASCADE,
+    product_id BIGINT NOT NULL REFERENCES products(product_id),
     quantity INT NOT NULL CHECK (quantity > 0),
     unit_price DECIMAL(10, 2) NOT NULL CHECK (unit_price >= 0),
     price_at_time DECIMAL(10, 2) NOT NULL CHECK (price_at_time >= 0)
